AIFoundation/AgentBuilder/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.6.1
langchain==0.0.335
langchain-openai==0.0.2
pinecone-client==2.2.4
pypdf2==3.0.1
python-dotenv==1.0.0
